requirements.txt
pyyaml==6.0.1
graphviz==0.20.1
pytest==7.4.3
